Move Align, Pack widgets to kas_core::hidden
