+++ crates/kas-widgets/src/adapt/mod.rs
@@ -9,7 +9,6 @@ mod adapt;
 mod adapt_cx;
 mod adapt_events;
 mod adapt_widget;
-mod align;
 mod reserve;
 mod with_label;
 
@@ -17,7 +16,7 @@ pub use adapt::{Adapt, Map};
 pub use adapt_cx::{AdaptConfigCx, AdaptEventCx};
 pub use adapt_events::AdaptEvents;
 pub use adapt_widget::*;
-pub use align::{Align, Pack};
 #[doc(inline)] pub use kas::hidden::MapAny;
+pub use kas::hidden::{Align, Pack};
 pub use reserve::{Margins, Reserve};
 pub use with_label::WithLabel;
